Extract allocations helper
