+++ test/benchmarks/bench_octicons.rb
@@ -4,6 +4,8 @@ require "minitest/benchmark"
 require "test_helper"
 
 class BenchOcticons < Minitest::Benchmark
+  include Primer::AssertAllocationsHelper
+
   def setup
     @options = {
       icon: :alert
@@ -13,7 +15,7 @@ class BenchOcticons < Minitest::Benchmark
   def bench_allocations_without_cache
     Primer::OcticonComponent.new(**@options)
     Primer::Octicon::Cache.clear!
-    assert_allocations 50 do
+    assert_allocations 51 do
       Primer::OcticonComponent.new(**@options)
     end
   ensure
@@ -22,22 +24,8 @@ class BenchOcticons < Minitest::Benchmark
 
   def bench_allocations_with_cache
     Primer::Octicon::Cache.preload!
-    assert_allocations 21 do
+    assert_allocations 19 do
       Primer::OcticonComponent.new(**@options)
     end
-  end
-
-  private
-
-  def assert_allocations(count)
-    GC.disable
-    total_start = GC.stat[:total_allocated_objects]
-    yield
-    total_end = GC.stat[:total_allocated_objects]
-    GC.enable
-
-    total = total_end - total_start
-
-    assert_equal count, total, "Expected between #{count} allocations, got #{total} allocations."
   end
 end
