+++ test/test_helper.rb
@@ -8,6 +8,7 @@ require "mocha/minitest"
 require "rails"
 require "rails/test_help"
 require "view_component/test_helpers"
+require "test_helpers/assert_allocations_helper"
 require "test_helpers/component_test_helpers"
 require "test_helpers/env_helper"
 require "yard/docs_helper"
